- Updated Scala versions - Added Scala reflexion library - Put the types ScalaEdge, ScalaGraph, ScalaVertex and ScalaElement in separate files for better readability - Moved package object scala in the package scala according to the convention - Method save() to save a class in the graph - Method load() to load a Vertex as an instance of class. This method is wrapped in an implicit object to import in order to be available only when dealing with Vertex (I got the idea from Apache Spark).
